requirements.txt
discord==1.7.3
python-dotenv==0.20.0

pysen==0.10.4
black==21.10b0
flake8==4.0.1
isort==5.10.1
mypy==0.910
pylint==2.13.9

# for vscode
pysen-ls>=0.1.0
